Various
Allows the setting of mins and/or maxes for validating user input as a number. Parameters now allow customizing whether or not "back" or "exit" is acceptable. Added new class for validating against regular expressions, either matching the full user input against the expression or searching for the expression within the user's input. Added new Exception class if the validate_user_input_num() method is called but arguments passed in preclude any allowed user input.